requirements.txt
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
django_rest_swagger==2.2.0
django-safedelete==0.5.7
drf-yasg==1.20.0